ai/requirements.txt
accelerate
diffusers
fastapi
imageio
moviepy
Pillow
python-multipart
torch
transformers
uvicorn[standard]
